pyproject.toml
[tool.poetry]
name = "ska-tmc-mid-integration"
version = "0.19.0"
description = "Repository for integration of TMC components"
authors = ["Team Himalaya", "Team Sahyadri"]
license = "BSD-3-Clause"

include = [
    { path = 'tests'}
]

[[tool.poetry.source]]
name = 'ska-nexus'
url = 'https://artefact.skao.int/repository/pypi-internal/simple'

[[tool.poetry.source]]
name = "PyPI-public"
url = 'https://pypi.org/simple'

[tool.poetry.dependencies]
python = "^3.10"
ska-control-model="0.3.1"
ska-tango-base="0.19.1"
pytango = "9.4.2"
numpy = "1.23.2"
ska-tango-testing = "0.6.0"
ska-ser-logging = "0.4.1+0a677a67"
httpx = "^0.19.0"
msgpack = "1.0.7"
msgpack_numpy = "0.4.8"
astropy = "^5.0.4"

[tool.poetry.dev-dependencies]
black = "^22.8.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
pylint = "^2.12.2"
pytest = "^7.1"
pytest-bdd = "^5.0.0"
pytest-cov = "^2.10.1"
pylint-junit = "^0.3.2"
coverage = "^6.2"
pytest-json-report = "^1.4.1"
pytest-forked = "^1.3.0"
pytest-mock = "^3.6.1"
pytest-xdist = "^2.4.0"
pytest-repeat = "^0.9.1"

[tool.poetry.group.docs.dependencies]
docutils = "^0.18.1"
Sphinx = "^5.1.1"
ska-ser-sphinx-theme = "^0.1.1"
sphinx-autodoc-typehints = "^1.19.3"
sphinxcontrib-plantuml = "^0.22"
typing_extensions  = "^4.3.0"
sphinx-rtd-theme = "^1.2.2"

[tool.flake8]
rst-directives = "literalinclude"
rst-roles = ["py:attr", "py:class", "py:const", "py:exc", "py:func", "py:meth", "py:mod"]
docstring-style = "sphinx"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
